Switch fps from int to float
It allows to have a perfect synchronization with cd music in kof98 intro, and
it shouldn't slow down anything noticeably normally !

+++ source/alleg/profile.c
@@ -9,7 +9,7 @@
 #include <sdl/SDL_gfx/SDL_framerate.h>
 #endif
 
-int fps = 60;
+float fps = 60.0;
 
 int use_rdtsc;
 UINT32 timer_next_update;	// timer_frame_count at next fps() update
